OJ-2714: Add ExcludeMethodFromGeneratedCoverage
Due to reason below, it is better to remove those private methods
and create the instances for VerifiableCredentialService in the
default constructor

Not really sure about this, it is fine in this context b'cos the
constructor in the handler is excluded from coverage. So it does make
sense to omit private method used within the context of a constructor
that has no coverage.

The downside to this is that future dev(s) could abuse this and decide to
omit methods that should have coverage by applying this annotations