OJ-2720: Use new cri-lib method to add kid to jwt header
Bump cri_common_lib 3.0.3 -> 3.0.4

Updated VerifiableCredentialService to use the new jwt signing method.

Updated IssueCredentialHandler to handle the NoSuchAlgorithmException. We opted to treat is as a 500 internal error. It is thrown from the new cri-lib method, but it would be very unusal to be thrown especially once pre-merge tests have ran. If this exception is ever thrown we may want to look at a differnt way to handle it.

Added and updated unit/pact tests for the above.