OJ-3037: Add POST postcode-lookup endpoint
To change the postcode-lookup endpoint to be a POST request with a body instead of a GET request with the postcode PII in the URL (issues with logging) I have updated the APIGW and existing PostcodeLookup lambda to handle both GET and POST requests.

Refactored lambda unit tests to include tests for both POST and GET. When GET is removed these tests will need removing and only the POST ones need to remain. The majority of tests on include POST requests.